Added backgrounds, shadows, color drivers, re-org
Dealt with restructuring the code to be a more pythonic package now before it gets too messy, each object is a separate file, which makes it easier to manage now that there are so many extra lines for:
Color drivers! I found a youtube video by Loop that showed that although you can only set float values as drivers, you can still target the individual channels of a color to the individual channels of a driver, so you CAN use colors as drivers to drive other colors! This means that the colors of the colorramps in each material can be assigned to custom properties and adjusted from the main viewport.
There was also a cool video by LanceBeryl.Dev who demonstrated that shadows could be added to Kristof Dedene's Grass material through the use of normals and a hue/saturation node. I have implemented this method on the grass, rock and water materials, and the shadows look very nice. A bit jaggy at times, but I will deal with that later.
Finally, added four different background shaders: Blue Sky, Sunset, Twilight and Starry Night. They are minimal, mostly just gradients (Starry Night also has stars...) that can be used in conjunction with a forthcoming cloud object.