Colored Light Support! Action Plane Drivers
Thanks to POLYCOSM pointing out that a translucent shader will work like an emission shader that accepts light information I have now reworked the grass, water and rock objects to receive colored light from regular lights in the scene, and cast shadows, and it even took fewer nodes!
The big trick is the same as the one shown by LanceBerylDev and others, but, instead of plugging a normal node > diffuse > shader to rgb > hue/saturation > mixrgb, we just remove all our previous shadow nodes, remove the emission node, replace it with the translucent node, plug the normal node into that, and invert the normal output Z value. That last step is critical.

This method comes with a big catch, though, and that is that light does not react with equal intensity across all colors.
Yellow light is the strongest, followed by green, followed by red, followed by blue. 
I think this is because we're pulling light data from normal data, and I've tried to correct with vector curves, colorramps, rgb curves, separateXYZ and RGB to tweak individual values, but so far they only shift the light location (as you would expect of a normal map), but they have no effect on the intensity of light across the color spectrum.
I am working around this by simply adjusting the intensity of individual lights, e.g. if the red lights seem less bright than the green I just raise the intensity of the red lights.
If I can find a solution to this I will, but for now I'm prioritizing adding objects and drivers.