Added Fire, Heat Ripple, Smoke Ring & Trail
Added FirePlane, HeatRipplePlane, Smoke Ring, Smoke Trail, updated the smoke material with a fresnel to better accentuate the curves of the clouds, updated smoke picture to match. Updated road material as well, after uploading it with misaligned roadlines. 
From the UI standpoint: renamed Effects to BG Effects, renamed Backgrounds to World Backgrounds, in anticipation of having to split the Objects category, and to help (hopefully) with clarity, to distinguish between something like the Energy Ring, which could be considered an effect, but takes place onscreen, and can be seen by the characters, and the Action Panels, which appear onscreen, but not in the "world" of the anime, and cannot be seen by other characters, but exist to enchance the action for the viewer.